pyproject.toml
[project]
name = "flowl"
version = "0.1.0"
description = "Probabilistic programming language based on optimal transport"
requires-python = ">=3.11"
dependencies = [
    "jax>=0.4.20",
    "jaxlib>=0.4.20",
    "ott-jax>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/flowl"]

[tool.pytest.ini_options]
testpaths = ["tests"]
